Update PlantWing projects

+++ Source/MeadowSamples/ConnectedPlant/PlantWing.App/PlantWing.App/Models/SoilMoisture.cs
@@ -6,7 +6,7 @@ namespace PlantWing.App.Models
     {
         public long Id { get; set; }
         public int Level { get; set; }
-        public int Moisture { get; set; }
+        public decimal Moisture { get; set; }
         public string Date { get; set; }
     }
 }
